requirements.txt
langchain>=0.0.335
langchain-core>=0.1.5
langchain-community>=0.0.1
langgraph>=0.0.25
fastapi>=0.103.1
uvicorn>=0.23.2
python-dotenv>=1.0.0
pandas>=1.0.0
google-generativeai>=0.3.0
